Adds AAPBCatalogSearch object
This object is designed to take in an AAPB search URL (cut and pasted from a browser) and
return a full list of JSON documents in the smallest number of queries (i.e. with page sizes
of 100 records per query). The resulting JSON objects can then be mapped to whatever subset
of properies you want, e.g. 'id', 'transcript_status', etc.